~Fixed struct type compilation
